small refactor to split up transfer service (#88)
